+++ CHANGELOG.md
@@ -1,5 +1,10 @@
-menuCreater CHANGELOG
+menuCreator CHANGELOG
 ==============================
+# 0.0.2
+
+Michael Dunton - Wiring up mongo, services and put recipe
+
+
 # 0.0.1
 
-Michael Dunton - midnight ramblings 
+Michael Dunton - midnight ramblings